[CheatMeal] 刪除作弊餐餐點 API

+++ PersonalTools/src/main/java/com/robbie/personaltools/infra/databases/dao/cheatmeal/MealDao.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MealDao extends JpaRepository<Meal, Integer> {
+public interface MealDao extends JpaRepository<Meal, Long> {
   Optional<Meal> findByName(String name);
 
   Page<Meal> findByNameContaining(String keyword, Pageable pageable);
